TINY-8279: moved most of the autocompleter to core
