+++ modules/tinymce/src/core/main/ts/alien/Spot.ts
