+++ modules/tinymce/src/core/main/ts/alien/TextDescent.ts
