+++ modules/tinymce/src/core/main/ts/alien/TextSearch.ts
